automod: quick extraction of setstore package.
Did not rename, so some names are stuttery (cachestore.CacheStore).
This can be fixed in a future pass if desired; today's goal is just
more packages and getting the filesystem tree approximately right.

I wanna log that I feel a little funky on this one.  I'm not fully
convinced that this actually needs an interface.  The main thing
that's expected to potentially change on this one is how it gets
loaded... but those methods _aren't in the interface anyway_.
The feature itself is providing data that's expected to be at low
(read: in-memory) latency: even if it was backed by redis as a data
source, the "load" function would be proactively forking that data
over into local memory.  A method on a concrete struct (so we're
still at least hiding direct access to the map) seems like it would
be sufficient.

But, all that said: maybe I'm wrong; there's very little harm in
having just one interface too many; and this is consistent with
the pattern of all the other things called "foostore".  So, fine.
